requirements_rpi5.txt
# ==============================
# Dépendances pour RPi5 + Hailo-8
# Installation : pip install -r requirements_rpi5.txt
# ==============================

# Core
numpy
opencv-python

# Object Detection (YOLO - utilisé en fallback si Hailo indisponible)
ultralytics

# Deep Learning Backend (version légère ARM64)
torch --index-url https://download.pytorch.org/whl/cpu
torchvision --index-url https://download.pytorch.org/whl/cpu

# Audio / Text-to-Speech
pyttsx3

# Espeak backend (requis par pyttsx3 sur Linux)
# sudo apt install espeak

# Hailo Runtime Python bindings
# Installé via apt (voir setup_rpi5.sh), pas via pip :
# sudo apt install hailo-all

# ==============================
# VLM — Moondream2
# ==============================
moondream
transformers
Pillow
einops
timm
